// src/main/frontend/src/types/groomingMenu.ts
export interface GroomingMenu {
    groomingNo: number; // 미용메뉴번호
    groomingName: string; // 미용메뉴이름
    groomingPrice: number; // 미용메뉴가격
    categoryNo: number; // 미용카테고리번호
    categoryName: string; // 미용카테고리이름
    storeNo: number; // 매장번호
}